requirements.txt
streamlit
whisper
openai-whisper
torch
qrcode
pillow
